move createTwo and createThree to test helpers
