+++ test/helpers/create.js
@@ -4,9 +4,40 @@ var ram = require('random-access-memory')
 module.exports = create
 
 create.one = createOne
+create.two = createTwo
+create.three = createThree
 
 function createOne (id) {
   return hyperdb(ram, {id: id, reduce: reduce, valueEncoding: 'json'})
+}
+
+function createTwo (cb) {
+  var a = hyperdb(ram, {valueEncoding: 'json'})
+  a.ready(function () {
+    var b = hyperdb(ram, a.key, {valueEncoding: 'json'})
+    b.ready(function () {
+      a.authorize(b.local.key, function () {
+        cb(a, b)
+      })
+    })
+  })
+}
+
+function createThree (cb) {
+  var a = hyperdb(ram, {valueEncoding: 'json'})
+  a.ready(function () {
+    var b = hyperdb(ram, a.key, {valueEncoding: 'json'})
+    b.ready(function () {
+      var c = hyperdb(ram, a.key, {valueEncoding: 'json'})
+      c.ready(function () {
+        a.authorize(b.local.key, function () {
+          b.authorize(c.local.key, function () {
+            cb(a, b, c)
+          })
+        })
+      })
+    })
+  })
 }
 
 function create (id) {
