+++ test/replicate.js
@@ -1,4 +1,5 @@
 var replicate = require('./helpers/replicate')
+var create = require('./helpers/create')
 var tape = require('tape')
 
 var hyperdb = require('..')
@@ -7,7 +8,7 @@ var ram = require('random-access-memory')
 tape('basic replication', function (t) {
   t.plan(6)
 
-  createTwo(function (a, b) {
+  create.two(function (a, b) {
     a.put('/a', 'a', function () {
       replicate(a, b, validate)
     })
@@ -31,7 +32,7 @@ tape('basic replication', function (t) {
 tape('2 peers, fork', function (t) {
   t.plan(20)
 
-  createTwo(function (a, b) {
+  create.two(function (a, b) {
     a.put('/root', 'root', function () {
       replicate(a, b, function () {
         a.put('/key', 'a', function () {
@@ -71,7 +72,7 @@ tape('2 peers, fork', function (t) {
 tape('2 peers, fork and non-merge write', function (t) {
   t.plan(20)
 
-  createTwo(function (a, b) {
+  create.two(function (a, b) {
     a.put('/root', 'root', function () {
       replicate(a, b, function () {
         a.put('/key', 'a', function () {
@@ -115,7 +116,7 @@ tape('2 peers, fork and non-merge write', function (t) {
 tape('2 peers, 1 reference old value', function (t) {
   t.plan(24)
 
-  createTwo(function (a, b) {
+  create.two(function (a, b) {
     a.put('/a', 'old', function () {
       replicate(a, b, function () {
         a.put('/a', 'new', function () {
@@ -163,7 +164,7 @@ tape('2 peers, 1 reference old value', function (t) {
 tape('2 peers, fork and merge write', function (t) {
   t.plan(16)
 
-  createTwo(function (a, b) {
+  create.two(function (a, b) {
     a.put('/root', 'root', function () {
       replicate(a, b, function () {
         a.put('/key', 'a', function () {
@@ -204,7 +205,7 @@ tape('2 peers, fork and merge write', function (t) {
 tape('3 peers', function (t) {
   t.plan(12)
 
-  createThree(function (a, b, c) {
+  create.three(function (a, b, c) {
     c.put('/test', 'test', function () {
       replicateThree(a, b, c, function () {
         a.get('/test', ontest)
@@ -225,7 +226,7 @@ tape('3 peers', function (t) {
 tape('3 peers + fork', function (t) {
   t.plan(18)
 
-  createThree(function (a, b, c) {
+  create.three(function (a, b, c) {
     a.put('/test', 'a', function () {
       c.put('/test', 'c', function () {
         replicateThree(a, b, c, function () {
@@ -298,35 +299,6 @@ function replicateThree (a, b, c, cb) {
     replicate(a, b, function (err) {
       if (err) return cb(err)
       replicate(b, c, cb)
-    })
-  })
-}
-
-function createTwo (cb) {
-  var a = hyperdb(ram, {valueEncoding: 'json'})
-  a.ready(function () {
-    var b = hyperdb(ram, a.key, {valueEncoding: 'json'})
-    b.ready(function () {
-      a.authorize(b.local.key, function () {
-        cb(a, b)
-      })
-    })
-  })
-}
-
-function createThree (cb) {
-  var a = hyperdb(ram, {valueEncoding: 'json'})
-  a.ready(function () {
-    var b = hyperdb(ram, a.key, {valueEncoding: 'json'})
-    b.ready(function () {
-      var c = hyperdb(ram, a.key, {valueEncoding: 'json'})
-      c.ready(function () {
-        a.authorize(b.local.key, function () {
-          b.authorize(c.local.key, function () {
-            cb(a, b, c)
-          })
-        })
-      })
     })
   })
 }
